chore: move fixtures to  tests/conftest.pyc
