+++ tests/test_options.py
@@ -24,23 +24,6 @@ from firebase_functions.private.serving import functions_as_yaml, merge_required
 # pylint: disable=protected-access
 
 
-@pytest.fixture(autouse=True)
-def _cleanup_params():
-    """Clear the global params registry so each test runs with a clean state."""
-    params._params.clear()
-    yield
-    params._params.clear()
-
-
-@pytest.fixture(autouse=True)
-def _cleanup_global_options():
-    """Reset global options so each test runs with a clean state."""
-    original_options = options._GLOBAL_OPTIONS
-    options._GLOBAL_OPTIONS = options.RuntimeOptions()
-    yield
-    options._GLOBAL_OPTIONS = original_options
-
-
 @https_fn.on_call()
 def asamplefunction(_):
     return "hello world"
